Fixing edge case on importing nginx logs
Because logrotate is separate from nginx, the log rotation may happen in
such a way that a log entry for the next day is emitted to the current
log file rather than the one that is being created for the new day. This
meant that when overwriting the aggregate stats, sometimes the last day
imported would be overwritten by the results of aggregating a single
request.

A new DAYS_BACK environment variable allows us to control how far back
we go. It defaults to 14 days, but I can set it and run it once with a
larger range to import the current backlog. When I found we were losing
some stats, I increased the log rotate limit to keep more days until I
fixed this.